feat: font now loading correctly in both on the dev server and storybook
