+++ .storybook/preview.tsx
@@ -1,9 +1,23 @@
 import type { Preview } from "@storybook/react"
+import localFont from "next/font/local"
 import React from "react"
 
 import "@/styles/globals.css"
 
-import FontsProvider from "../lib/providers/FontsProvider"
+const GTFlexa = localFont({
+  src: [
+    {
+      path: "../fonts/GT-Flexa-Expanded-Regular.woff2",
+      weight: "400",
+    },
+    {
+      path: "../fonts/GT-Flexa-Expanded-Medium.woff2",
+      weight: "500",
+    },
+  ],
+  variable: "--font-headline",
+  display: "swap",
+})
 
 const preview: Preview = {
   parameters: {
@@ -21,8 +35,7 @@ const preview: Preview = {
     // 👇 Defining the decorator in the preview file applies it to all stories
     (Story, { parameters }) => {
       return (
-        <div>
-          <FontsProvider />
+        <div className={`${GTFlexa.variable} antialiased`}>
           <Story {...parameters} />
         </div>
       )
