+++ app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next"
+import localFont from "next/font/local"
 
 import Footer from "@/components/global/footer/Footer"
 import Header from "@/components/global/header/Header"
 import Theme from "@/components/global/theme/Theme"
-import FontsProvider from "@/lib/providers/FontsProvider"
 import ReactQueryProvider from "@/lib/providers/ReactQueryProvider"
 import "@/styles/globals.css"
 
@@ -12,6 +12,21 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 }
 
+const GTFlexa = localFont({
+  src: [
+    {
+      path: "../public/fonts/GT-Flexa-Expanded-Regular.woff2",
+      weight: "400",
+    },
+    {
+      path: "../public/fonts/GT-Flexa-Expanded-Medium.woff2",
+      weight: "500",
+    },
+  ],
+  variable: "--font-headline",
+  display: "swap",
+})
+
 export default function RootLayout({
   children,
 }: Readonly<{
@@ -19,8 +34,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className="antialiased">
-        <FontsProvider />
+      <body className={`${GTFlexa.variable} antialiased`}>
         <Theme>
           <ReactQueryProvider>
             <Header />
