+++ package.json
@@ -56,7 +56,7 @@
     "@storybook/addon-links": "^8.3.6",
     "@storybook/addon-onboarding": "^8.3.6",
     "@storybook/blocks": "^8.3.6",
-    "@storybook/nextjs": "^8.3.6",
+    "@storybook/nextjs": "^8.4.0",
     "@storybook/react": "^8.3.6",
     "@storybook/test": "^8.3.6",
     "@svgr/webpack": "^8.1.0",
@@ -80,7 +80,7 @@
     "prettier-plugin-classnames": "^0.7.3",
     "prettier-plugin-merge": "^0.7.1",
     "prettier-plugin-tailwindcss": "^0.6.8",
-    "storybook": "^8.3.6",
+    "storybook": "^8.4.0",
     "tailwindcss": "^3.4.13",
     "typescript": "^5.6.2",
     "vite-tsconfig-paths": "^5.0.1",
