+++ public/fonts/GT-Flexa-Expanded-Regular.woff2
